+++ framework/cmsunittests/InitialPopulationTests.cs
@@ -15,10 +15,10 @@ namespace cmsunittests
             Configuration.CurrentConfiguration = Configuration.ConfigurationFromString("{}");
             RNGFactory.Reset();
             var modelInfo = compartments.emodl.EmodlLoader.LoadEMODLModel("(import (rnrs) (emodl cmslib)) (start-model \"fixedPopulation\") (species S 42) (species I 1) (observe susceptible S) (reaction infect (S) (I) (* S I)) (end-model)");
-            var results = compartments.Program.ExecuteModel(modelInfo, "SSA", 10.0f, 1000, 4);
+            var results = compartments.Program.ExecuteModel(modelInfo, "SSA", 10.0, 1000, 4);
             foreach (var vector in results.Data)
             {
-                Assert.AreEqual(42.0f, vector[0]);
+                Assert.AreEqual(42.0, vector[0]);
             }
         }
 
@@ -26,25 +26,25 @@ namespace cmsunittests
         public void UniformPopulationTest()
         {
             const string distribution = "uniform 21 64";
-            float mean;
-            float stdDev;
+            double mean;
+            double stdDev;
             RunPopulationTest(distribution, out mean, out stdDev);
-            Assert.IsTrue((40.0f <= mean) && (mean <= 44.0f));
-            Assert.IsTrue((11.0f <= stdDev) && (stdDev <= 14.0f));
+            Assert.IsTrue((40.0 <= mean) && (mean <= 44.0));
+            Assert.IsTrue((11.0 <= stdDev) && (stdDev <= 14.0));
         }
 
-        private static void RunPopulationTest(string distribution, out float mean, out float stdDev)
+        private static void RunPopulationTest(string distribution, out double mean, out double stdDev)
         {
             Configuration.CurrentConfiguration = Configuration.ConfigurationFromString(string.Format("{{\"prng_seed\":{0},\"prng_index\":{1}}}", DateTime.Now.Millisecond, DateTime.Now.DayOfYear));
             RNGFactory.Reset();
             const string modelTemplate = "(import (rnrs) (emodl cmslib)) (start-model \"fixedPopulation\") (species S ({0})) (species I 1) (observe susceptible S) (reaction infect (S) (I) (* S I)) (end-model)";
             var modelInfo = compartments.emodl.EmodlLoader.LoadEMODLModel(string.Format(modelTemplate, distribution));
-            var results = compartments.Program.ExecuteModel(modelInfo, "SSA", 10.0f, 1000, 4);
-            float sum = results.Data.Sum(vector => vector[0]);
-            float sumSqr = results.Data.Sum(vector => (vector[0]*vector[0]));
-            float n = results.Data.Length;
+            var results = compartments.Program.ExecuteModel(modelInfo, "SSA", 10.0, 1000, 4);
+            double sum = results.Data.Sum(vector => vector[0]);
+            double sumSqr = results.Data.Sum(vector => (vector[0]*vector[0]));
+            double n = results.Data.Length;
             mean = sum/n;
-            stdDev = (float) Math.Sqrt((sumSqr - 2*mean*sum + n*mean*mean)/n);
+            stdDev = Math.Sqrt((sumSqr - 2*mean*sum + n*mean*mean)/n);
             Console.WriteLine("initial population (mean):      {0}", mean);
             Console.WriteLine("initial population (std. dev.): {0}", stdDev);
         }
@@ -53,11 +53,11 @@ namespace cmsunittests
         public void NormalPopulationTest()
         {
             const string distribution = "normal 42 4";
-            float mean;
-            float stdDev;
+            double mean;
+            double stdDev;
             RunPopulationTest(distribution, out mean, out stdDev);
-            Assert.IsTrue((40.0f <= mean) && (mean <= 44.0f));
-            Assert.IsTrue((3.5f <= stdDev) && (stdDev <= 4.5f));
+            Assert.IsTrue((40.0 <= mean) && (mean <= 44.0));
+            Assert.IsTrue((3.5 <= stdDev) && (stdDev <= 4.5));
         }
     }
 }
